move code from Program into SourceConverter
